requirements.txt
numpy==1.26.4
torch==2.2.1
torchvision==0.17.1
transformers==4.38.2
pillow==10.2.0
requests==2.31.0